Generalize Tool implementation.
