+++ src/article/editor/InsertInlineGraphicTool.js
@@ -2,7 +2,7 @@ import UploadTool from '../shared/UploadTool'
 
 export default class InsertInlineGraphicTool extends UploadTool {
   getClassNames () {
-    return 'sc-insert-inline-graphic-tool sm-upload-tool'
+    return 'sc-insert-inline-graphic-tool sc-upload-tool sc-tool'
   }
 
   getFileType () {
