+++ src/article/editor/SupplementaryFileTools.js
@@ -12,12 +12,12 @@ class SupplementaryFileTool extends UploadTool {
 
 export class InsertSupplementaryFileTool extends SupplementaryFileTool {
   getClassNames () {
-    return 'sc-insert-supplementary-file-tool sm-upload-tool'
+    return 'sc-insert-supplementary-file-tool sc-upload-tool'
   }
 }
 
 export class ReplaceSupplementaryFileTool extends SupplementaryFileTool {
   getClassNames () {
-    return 'sc-replace-supplementary-file-tool sm-upload-tool'
+    return 'sc-replace-supplementary-file-tool sc-upload-tool'
   }
 }
