+++ src/article/editor/styles/_upload-tool.css
@@ -1,3 +1,3 @@
-.sm-upload-tool > input {
+.sc-upload-tool > input {
   display: none;
 }
