+++ src/kit/ui/Menu.js
@@ -14,14 +14,14 @@ export default class Menu extends Component {
     for (let item of items) {
       if (item.type === 'command') {
         const commandState = commandStates[item.name] || DISABLED
-        const ItemComponent = item.component ? this.getComponent(item.component) : this.getComponent('menu-item')
+        const ToolClass = this._getToolClass(item)
         el.append(
-          $$(ItemComponent, {
+          $$(ToolClass, {
             style,
             theme,
             item,
             commandState
-          })
+          }).ref(item.name).addClass('se-item')
         )
       } else if (item.type === 'separator') {
         el.append(
@@ -32,5 +32,23 @@ export default class Menu extends Component {
       }
     }
     return el
+  }
+
+  _getToolClass (item) {
+    // use an ToolClass from toolSpec if configured inline in ToolGroup spec
+    let ToolClass = item.ToolClass
+    // next try if there is a tool registered by the name
+    if (!ToolClass) {
+      ToolClass = this.context.toolRegistry.get(item.name)
+    }
+    // after all fall back to default classes
+    if (!ToolClass) {
+      if (this.props.style === 'descriptive') {
+        ToolClass = this.getComponent('menu-item')
+      } else {
+        ToolClass = this.getComponent('toggle-tool')
+      }
+    }
+    return ToolClass
   }
 }
