+++ src/kit/ui/index.js
@@ -41,6 +41,7 @@ export { default as TextNodeComponent } from '../ui/TextNodeComponent'
 export { default as TextPropertyComponent } from './_TextPropertyComponent'
 export { default as TextPropertyEditor } from './_TextPropertyEditor'
 export { default as ToggleTool } from './ToggleTool'
+export { default as Tool } from './Tool'
 export { default as Toolbar } from './Toolbar'
 export { default as ToolDropdown } from './ToolDropdown'
 export { default as ToolGroup } from './ToolGroup'
